Ouioui baguette

+++ crates/terrain/src/mesher.rs
@@ -65,7 +65,7 @@ pub fn generate_model(voxels: &Box<[Voxel]>, size: usize, interpolation: bool, s
                         let voxel1 = voxels[index1];
                         let voxel2 = voxels[index2];
                         // Do inverse linear interpolation to find the factor value
-                        let value: f32 = if interpolation { inverse_lerp(voxel1.density as f32, voxel2.density as f32, ISOLINE as f32) } else { 0.5 };
+                        let value: f32 = if interpolation { inverse_lerp(voxel1.density, voxel2.density, ISOLINE as f32) } else { 0.5 };
                         // Create the vertex
                         let mut vertex = veclib::Vector3::<f32>::lerp(vert1, vert2, value);
                         // Offset the vertex
